Update CustomLayouts after saving changes in a Class

+++ models/DataObject/ClassDefinition.php
@@ -404,8 +404,11 @@ final class ClassDefinition extends Model\AbstractModel implements ClassDefiniti
         } else {
             $this->dispatchEvent(new ClassDefinitionEvent($this), DataObjectClassDefinitionEvents::POST_ADD);
         }
-
-        $this->deleteDeletedDataComponentsInCustomLayout();
+        if(!empty($this->getDeletedDataComponents())) {
+            $this->deleteDeletedDataComponentsInCustomLayout();
+        } else {
+            $this->updateCustomLayouts();
+        }
     }
 
     /**
@@ -444,7 +447,7 @@ final class ClassDefinition extends Model\AbstractModel implements ClassDefiniti
             \Pimcore\File::putPhpFile($definitionFile, $data);
         }
     }
-
+    
     /**
      * @internal
      */
@@ -1138,11 +1141,22 @@ final class ClassDefinition extends Model\AbstractModel implements ClassDefiniti
         return $this;
     }
 
+    private function updateCustomLayouts(): void {
+        
+        $customLayouts = new ClassDefinition\CustomLayout\Listing();
+        $id = $this->getId();
+        $customLayouts->setFilter(function (DataObject\ClassDefinition\CustomLayout $layout) use ($id) {
+            return $layout->getClassId() === $id;
+        });
+        $customLayouts = $customLayouts->load();
+
+        foreach ($customLayouts as $customLayout) {
+            $customLayout->save();
+        }
+    }
+
     private function deleteDeletedDataComponentsInCustomLayout(): void
     {
-        if (empty($this->getDeletedDataComponents())) {
-            return;
-        }
         $customLayouts = new ClassDefinition\CustomLayout\Listing();
         $id = $this->getId();
         $customLayouts->setFilter(function (DataObject\ClassDefinition\CustomLayout $layout) use ($id) {
